requirements.txt
asgiref==3.11.1
Django==6.0.3
sqlparse==0.5.5
Pillow>=10.0.0
django-modeltranslation>=0.18
